Remove all Godeps dependencies; fully convert to dep
The project now uses [dep][] for dependency management. The `vendor/` folder is excluded from source control so after a fresh clone, you must run `dep ensure -v` to _ensure_ (see what I did there?!) that all the dependencies are downloaded and synced to the appropriate version before building.

And yeah, the `-v` option is optional, but I strongly recommend it. The dep commands can be very slow, and there is absolutely zero user feedback without the verbose (`-v`) flag. It's bad UX.

Of course all of this also means you need to get dep itself. :) The easiest way to do that is to grab a binary directly from their [releases][] page. It's just a standalone executable that you download and run, or add to your `PATH`, etc. That's the beauty of Go! :) There's more information, including additional installation methods, in their [installation guide][], too.

[dep]: https://golang.github.io/dep/docs/introduction.html
[releases]: https://github.com/golang/dep/releases
[installation guide]: https://golang.github.io/dep/docs/installation.html